pyproject.toml
[project]
name = "hackathon-5"
version = "0.1.0"
description = "FlowForge Customer Success Digital FTE — Production AI Agent"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.2.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.10.0",
    "groq>=0.15.0",
    "openai-agents>=0.13.5",
    "httpx>=0.28.1",
    "twilio>=9.10.4",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "aiokafka>=0.13.0",
    "apscheduler>=3.11.2",
]
